==> Part 1 - Content Based Recommender - Exploration and Cleaning.ipynb <==
import pandas as pd

#### Movies

movies_df = pd.read_csv('ml-20m/movies.csv')

# provides information movieId, title, and genres of movies
# genres will need to be split up into seperate features for each individual movie
movies_df.head()

movies_df.shape

movies_df.isnull().sum()

movies_df.info()

movies_df.describe().T

    MIGHT NEED TO REMOVE SINCE IT'S THE OTHER FILE OR ADD TO THE MOVIES DATASET

#### Ratings 

ratings_df = pd.read_csv('ml-20m/ratings.csv')

ratings_df.head()

ratings_df.isnull().sum()

ratings_df.info()

ratings_df.shape

ratings_df = ratings_df.drop(columns='timestamp')

ratings_df.describe().T

# Create and export the average rating and number of ratings by movie for EDA
avg_rating = ratings_df.groupby('movieId').mean()
num_rating = ratings_df.groupby('movieId').count()

# Drop unnecessary solumns and rename the rating column
avg_rating = avg_rating.rename(columns={'rating':'avg_rating'})
avg_rating = avg_rating.drop(columns= 'userId')
avg_rating['movieId'] = avg_rating.index
avg_rating.head(3)

num_rating = num_rating.rename(columns={'rating':'num_rating'})
num_rating = num_rating.drop(columns= 'userId')
num_rating['num_rating'] = num_rating.index
num_rating.head(3)

rating_eda = pd.merge(avg_rating, num_rating, 
                      left_on = avg_rating['movieId'], 
                      right_on= num_rating.index).drop(columns='key_0')
rating_eda.head()

rating_eda.shape

rating_eda.to_csv('data/rating_eda.csv',index = False)

#### Tags

tags_df = pd.read_csv('ml-20m/tags.csv')

tags_df.head()

tags_df.shape

tags_df.isnull().sum()

tags_df[tags_df['tag'].isnull()]

# We are going to replace the null values with "missing"
tags_df = tags_df['tag'].fillna('missing')

#### Links

links_df = pd.read_csv('ml-20m/links.csv')

# links between other tables will be necessary for joins later

# imdb id for actor / actress data
# imdbpy 

links_df.head()

links_df.shape

#### Genome Tags & Scores

genome_scores_df = pd.read_csv('ml-20m/genome-scores.csv')

genome_scores_df.head()

genome_scores_df.shape

genome_scores_df.isnull().sum()

genome_scores_df.info()

genome_tags_df = pd.read_csv('ml-20m/genome-tags.csv')

genome_tags_df.head()

genome_tags_df.shape

genome_tags_df.isnull().sum()

genome_tags_df.info()

# Combining the two datasets to get be able to merge tags with movies
genome_df = pd.merge(genome_tags_df, genome_scores_df, left_on = 'tagId', right_on= 'tagId')

# this provides each movie with grat amount of details
genome_df.sort_values(by = 'movieId').head()

# Exporting genome data for the recommender and eda
genome_df.to_csv('data/genome_clean.csv', index = False)

## Movie Cleanings

    Creating features from movies dataframe

movies_df.head(5)

movies_df['year'] = movies_df['title'].map(lambda x : x.split('(')[-1][:4])#.astype(int)

# look up all the years that were missing in the format
change_needed = []
for i in movies_df['year']:
    try:
        x = int(i)
    except:
        change_needed.append(i)
change_needed

# print the titles of the movies with missing years
for i in movies_df['year']:
    if i in change_needed:
        print(movies_df[movies_df['year']==i][['title','year']])

movies_df[movies_df['year']=='Baby'] = 1993
movies_df[movies_df['year']=='Das '] = 1970
movies_df[movies_df['year']=='Bici'] = 2010
movies_df[movies_df['year']=='Braz'] = 2014
movies_df[movies_df['year']=='Slay'] = 2014
movies_df[movies_df['year']=='Tato'] = 2014
movies_df[movies_df['year']=='Nati'] = 2011
movies_df[movies_df['year']=='The '] = 1990
movies_df[movies_df['year']=='In O'] = 2002
movies_df[movies_df['year']=='Step'] = 2008
movies_df[movies_df['year']=='Two:'] = 2008
movies_df[movies_df['year']=="Li'l"] = 2014
movies_df[movies_df['year']=='A Ye'] = 1991
movies_df[movies_df['year']=='Body'] = 1993
movies_df[movies_df['year']=='Pols'] = 2014
movies_df[movies_df['year']=='The '] = 1993
movies_df[movies_df['year']=='My O'] = 2014
movies_df[movies_df['year']=='Movi'] = 2003
movies_df[movies_df['year']=='Mich'] = 1993

# Converting  
movies_df['year'] = movies_df['year'].astype(int)

movies_df.to_csv('data/movies_cleaned.csv', index = False)

## User Reviews Based Recommendations

user based -- looking for movies that other users have watched that you have not yet: users as rows and movies as columns: based on your profile you will like this


item based -- item with similar user movies rows and users as columns: when you go to a movie or book and the remmendation there is normally item based. people who bought this also bought this

movies rows and users as columns

#ratings_df.head()

#user_based_df = pd.pivot_table(ratings_df, values = 'rating' , index = 'userId', columns = 'movieId')

#movie_based_df = pd.pivot_table(ratings_df, values = 'rating' , index = 'movieId', columns = 'userId')
