==> Part 2 - Content Based Recommender - Cleaning II.ipynb <==
import pandas as pd
from scipy import sparse
from sklearn.metrics.pairwise import pairwise_distances
from sklearn.feature_extraction.text import CountVectorizer

 - Movies Genres Feature Enginnering 
 - Genome Tranformations 
 - Combining the data
 - Recommendation matrix

# Reading in the features 
movies_df1 = pd.read_csv('data/movies_cleaned.csv', low_memory = False)
genome_df = pd.read_csv('data/genome_clean.csv', low_memory = False)
movies_df1.head(2)

genome_df.head(2)

# Creating a function to clean up genres and get the data ready for word vecotrizer
def no_genres(i):
    if i == '(no genres listed)':
        return 'no_genres'
    else:
        return i

movies_df1['genres'] = movies_df1['genres'].apply(no_genres)
movies_df1[movies_df1['genres'].str.contains('no')].head(3)

movies_df1.head(3)

#### Movies Genres Feature Enginnering 

# Create a decade feature for every movie, instead of a year

# Function to create take a floor round 
def to_decade(num):
    num_str = str(num)[:3] + '0'
    return num_str

# Assign the new feature
movies_df1['decade'] = movies_df1['year'].map(to_decade)
movies_df1.head(3)

# Creating dummy variables from decade
movies_df = pd.get_dummies(movies_df1, columns=['decade'], drop_first=False)
movies_df.head(3)

# Remove year from the dataframe, as it is not unnecessary
movies_df = movies_df.drop(columns='year',axis = 1 )

# Create features based each genre 

# instantiating count vectorizer
vectorizer = CountVectorizer(analyzer = "word",   
                             tokenizer = None,    
                             preprocessor = None, 
                             stop_words = None) 

# applying count vectorizer to yelp categories column
cat_vector = vectorizer.fit_transform(movies_df['genres']).toarray()

# merging results with original yelp data frame
cat_vector_df = pd.DataFrame(cat_vector, columns = vectorizer.get_feature_names())
cat_vector_df.shape

# Merging the vectorizer genres and the movie dataframes
movies_vecrt_df = pd.merge(left = movies_df, 
                            right = cat_vector_df, 
                            on=movies_df.index)

movies_vecrt_df.head(3)

# dropping the key_0 since it's unnecessary
movies_vecrt_df = movies_vecrt_df.drop(columns='key_0')

movies_vecrt_df.columns

# Rename address scf-fi movies as the category was split up

movies_vecrt_df = movies_vecrt_df.drop(columns = 'fi', axis = 1)
movies_vecrt_df = movies_vecrt_df.rename(columns = {'sci': 'scifi'})

# Rename address scf-fi movies as the category was split up

movies_vecrt_df = movies_vecrt_df.drop(columns = 'film', axis = 1)
movies_vecrt_df = movies_vecrt_df.rename(columns = {'noir': 'film_noir'})

movies_vecrt_df.head(3)

movies_vecrt_df['decade'] = movies_df1['decade']

# Exporting dataframe for data exploration
movies_vecrt_df.to_csv('data/movie_features_eda.csv',index = False)

movies_vecrt_df = movies_vecrt_df.drop(columns=['genres','decade'])

### Genome Tranformations 

genome_df.head(2)

genome_df.info()

genome_pivot_df = pd.pivot_table(genome_df, values = 'relevance' , index = 'movieId', columns = 'tag')

genome_pivot_df.head(3)

# left joining the genome features to the movies dataframe
df = pd.merge(left = movies_vecrt_df, 
                right = genome_pivot_df, 
                left_on=movies_df['movieId'], 
                right_on = genome_pivot_df.index,
                how = 'left').drop('key_0', axis = 1)
df.head(3)

# resetting the index to title and removing columns that are not features of the movie
df.index = df['title']
df = df.drop(columns = ['title','movieId'])
df.head()

### Creating Recommendation Matrix

# transforming the data to sparse matrix
df_sparse = sparse.csr_matrix(df.fillna(0))

recommender = pairwise_distances(df_sparse, metric = 'cosine')

# Checking the shape to see the alignment of movies with other movies
recommender.shape

# Creating a dataframe for the recommender
recommender_df = pd.DataFrame(recommender, index = df.index, columns = df.index)
recommender_df.head()

# Creating a formula to search through the recommendation engine and provide the top 10 picks

def top_ten_recommendations(search):
    for title in movies_df[movies_df['title'].str.lower().str.contains(search.lower())]['title'].values:
        print(title)
#         print('Average Rating', df[df['title'] == title]['rating'].mean())
#         print('Number of Ratings', df[df['title']==title].shape[0])
        print(" ")
        print(recommender_df[title].sort_values()[1:11])
        print(" ")
        print(" ")

top_ten_recommendations('elf')
