==> Part 6 - Collaborative Based Recommender - Cleaning Data.ipynb <==
import pandas as pd

### Read in and explore the the ratings data

# Reading in the rating file
with open('ml-10M100K/ratings.dat','r') as f:
    next(f) # skip first row
    df = pd.DataFrame(l.rstrip().split('::') for l in f)

df.head(3)

# Dropping the time stamp 
df = df.drop(columns = 3)

df.columns

# renaming columns
df = df.rename({0:'user_id', 1:'movie_id', 2:'rating'}, axis = 1)

df.head(3)

# Checking for nulls 
df.isnull().sum()

# Checking the size of our dataset
df.shape

# checking datatypes
df.info()

# Changing rating to float and user_id/movie_id to an integer 
df['rating'] = df['rating'].astype(float)
df['user_id'] = df['user_id'].astype(int)
df['movie_id'] = df['movie_id'].astype(int)

df.info()

# Checking the distribution of ratings
df.describe().T

    Due to the size of the file we are going to drop half the users 

df = df[df['user_id']%3==0]

df.shape

### Read the movies data and join it with our rating data

movies_df = pd.read_csv('ml-10M100K/movies.csv')

movies_df.head(3)

movies_df.shape

# drop the genre since I will not be using that columns
movies_df = movies_df.drop(columns='genres')

movies_df.info()

# Changing the data type of movie id to int for a join with the ratings dataset
movies_df['movieId'] = movies_df['movieId'].astype(int)

### Joining the rating and movies

df.info()

# Left joining the movies to the the rating data
final_df = pd.merge(df, movies_df, 
                    left_on = 'movie_id',
                    right_on = 'movieId',
                    how = 'left')

final_df.head()

# Dropping the movie_ids as they are no longer necessary
final_df = final_df.drop(columns=['movie_id','movieId'])

# double checking final shape
print(df.shape)
print(final_df.shape)

# Checking for null values
final_df.isnull().sum() 

# I am going to drop the N/As since it's a small percentage of our dataset
final_df = final_df.dropna()

final_df.head(3)

## User Reviews Based Recommendations

user based -- looking for movies that other users have watched that you have not yet: users as rows and movies as columns: based on your profile you will like this


item based -- item with similar user movies rows and users as columns: when you go to a movie or book and the remmendation there is normally item based. people who bought this also bought this

item_based_df = pd.pivot_table(final_df, values = 'rating' , index = 'title', columns = 'user_id')

item_based_df.head()

    Exporting the data for other notebooks

item_based_df.to_csv('data/item_based.csv')

user_based_df = pd.pivot_table(final_df, values = 'rating' , index = 'user_id', columns = 'title')
user_based_df.head(3)

user_based_df.to_csv('data/movie_based.csv')
