==> Part 8 - Collaborative Based Recommender - Users.ipynb <==
import pandas as pd
from scipy import sparse
from sklearn.metrics.pairwise import pairwise_distances
from sklearn.feature_extraction.text import CountVectorizer

### Creating User Based Recommendor

# we need to pull users and build the similarity matrix
# then look at top other users that are the most similar 
# check the movies that they have rated highly and this person has not seen

# Read and check our data
df = pd.read_csv('data/item_based.csv', index_col = 'title')
movies_df = pd.read_csv('ml-10M100K/movies.csv')
df.head()

# Transforming the data to sparse matrix
df_sparse = sparse.csr_matrix(df.fillna(0))

# Creating a recommender 
recommender = pairwise_distances(df_sparse, metric = 'cosine')

# Checking the shape of our recommender
recommender.shape

# Creating a dataframe for the recommender
recommender_df = pd.DataFrame(recommender, index = df.index, columns = df.index)
recommender_df.head(3)

# Creating a formula to search through the recommendation engine and provide the top 10 picks

def top_ten_recommendations(search):
    for title in movies_df[movies_df['title'].str.lower().str.contains(search.lower())]['title'].values:
        if title in recommender_df.index:
            print(title)
    #         print('Average Rating', df[df['title'] == title]['rating'].mean())
    #         print('Number of Ratings', df[df['title']==title].shape[0])
            print(" ")
            print(recommender_df[title].sort_values()[1:11])
            print(" ")
            print(" ")
#         else: 
#             print('Unfortunately there is this movie, please pick a new one.')

top_ten_recommendations('napoleon')
